Implemented everything except unit testing

+++ src/main/java/com/hexaware/maverickBank/repository/ICustomerRepository.java
@@ -1,10 +1,10 @@
 package com.hexaware.maverickBank.repository;
 
-import com.hexaware.maverickBank.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.hexaware.maverickBank.entity.Customer;
 @Repository
-public interface ICustomerRepository extends JpaRepository<Customer, Integer> {
-    Customer findByUser_UserId(int userId);
+public interface ICustomerRepository extends JpaRepository<Customer, Long> {
+    Customer findByUser_UserId(Long userId);
 }
